[AITO-280] refactor client, requests, and responses as client module
