+++ aito/client/responses/__init__.py
@@ -1,0 +1,8 @@
+"""Response classes returned by the :class:`~aito.client.AitoClient`
+
+"""
+from .aito_response import BaseResponse, GetVersionResponse
+from .job_api_response import CreateJobResponse, GetJobStatusResponse
+from .query_api_response import BaseHit, ScoredHit, ProbabilityHit, RelateHit, HitsResponse, SearchResponse, \
+    PredictResponse, RecommendResponse, SimilarityResponse, MatchResponse, RelateResponse, EvaluateResponse
+from .schema_api_response import DatabaseSchemaResponse, TableSchemaResponse, ColumnSchemaResponse
