+++ aito/client/responses/query_api_response.py
@@ -1,15 +1,8 @@
-"""Response objects returned by the :class:`~aito.client.AitoClient`
-
-"""
-
-import logging
 from abc import ABC, abstractmethod
-from typing import List, Dict, Any, Type, TypeVar, Generic, Optional
+from typing import TypeVar, Generic, Type, List, Dict, Any
 
-from aito.schema import AitoSchema, AitoDatabaseSchema, AitoTableSchema, AitoColumnTypeSchema
+from .aito_response import BaseResponse
 from aito.utils._json_format import JsonFormat
-
-LOG = logging.getLogger('AitoResponse')
 
 
 class BaseHit(JsonFormat):
@@ -122,52 +115,6 @@ class RelateHit(BaseHit):
     def probabilities(self):
         """return probabilities information of the relation"""
         return self.__getitem__('ps')
-
-
-class BaseResponse(JsonFormat):
-    """The base class for the AitoClient request response
-
-    """
-    def __init__(self, json: Dict):
-        """
-        :param json: the original JSON response of the request
-        :type json: Dict
-        """
-        self.json_schema_validate(json)
-        self._json = json
-
-    @property
-    def json(self):
-        """the original JSON response of the request
-
-        :rtype: Dict
-        """
-        return self._json
-
-    def __getitem__(self, item):
-        if item not in self._json:
-            raise KeyError(f'Response does not contain field `{item}`')
-        return self._json[item]
-
-    def __contains__(self, item):
-        return item in self._json
-
-    def __iter__(self):
-        return iter(self._json)
-
-    def __len__(self):
-        return len(self._json)
-
-    @classmethod
-    def json_schema(cls):
-        return {'type': 'object'}
-
-    def to_json_serializable(self):
-        return self._json
-
-    @classmethod
-    def from_deserialized_object(cls, obj: Any):
-        return cls(obj)
 
 
 HitType = TypeVar('HitType', bound=BaseHit)
@@ -393,94 +340,3 @@ class EvaluateResponse(BaseResponse):
         :rtype: int
         """
         return self.__getitem__('trainSamples')
-
-
-class GetVersionResponse(BaseResponse):
-    """Response of the get version request"""
-    @property
-    def version(self) -> str:
-        """the Aito instance version
-
-        :rtype: str
-        """
-        return self.__getitem__('version')
-
-
-class _SchemaResponse(BaseResponse, ABC):
-    """Response that contains a schema component"""
-    @property
-    @abstractmethod
-    def schema_cls(self) -> Type[AitoSchema]:
-        """the class of the schema component
-
-        :rtype: Type[AitoSchema]
-        """
-        pass
-
-    @property
-    def schema(self) -> AitoSchema:
-        """return an instance of the appropriate AitoSchema
-
-        :rtype: AitoSchema
-        """
-        return self.schema_cls.from_deserialized_object(self._json)
-
-
-class DatabaseSchemaResponse(_SchemaResponse):
-    """Response that contains a database schema"""
-    @property
-    def schema_cls(self) -> Type[AitoSchema]:
-        return AitoDatabaseSchema
-
-
-class TableSchemaResponse(_SchemaResponse):
-    """Response that contains a table schema"""
-    @property
-    def schema_cls(self) -> Type[AitoSchema]:
-        return AitoTableSchema
-
-
-class ColumnSchemaResponse(_SchemaResponse):
-    """Response that contains a column schema"""
-    @property
-    def schema_cls(self) -> Type[AitoSchema]:
-        return AitoColumnTypeSchema
-
-
-class CreateJobResponse(BaseResponse):
-    """Response of the `Create job <https://aito.ai/docs/api/#post-api-v1-jobs-query>`__
-    containing the job information
-    """
-    @property
-    def id(self) -> str:
-        """the id of a job"""
-        return self.__getitem__('id')
-
-    @property
-    def started_at(self) -> str:
-        """when the job was started"""
-        return self.__getitem__('startedAt')
-
-
-class GetJobStatusResponse(BaseResponse):
-    """Response of the `Get job status <https://aito.ai/docs/api/#get-api-v1-jobs-uuid>`__"""
-
-    @property
-    def id(self) -> str:
-        """the id of the job session"""
-        return self.__getitem__('id')
-
-    @property
-    def started_at(self) -> str:
-        """when the job was started"""
-        return self.__getitem__('startedAt')
-
-    @property
-    def finished(self) -> bool:
-        """if the job has finished"""
-        return self.__contains__('finishedAt')
-
-    @property
-    def expires_at(self) -> Optional[str]:
-        """when the job result will not be available, returns None if job hasn't finished"""
-        return self.__getitem__('startedAt') if self.finished else None
