+++ tests/inline_docs/test_client_inline_docs.py
@@ -6,11 +6,11 @@ import aito.client
 
 
 def load_tests(loader, tests, ignore):
-    grocery_demo_client = aito.client.AitoClient(
+    grocery_demo_client = aito.client.aito_client.AitoClient(
         environ['AITO_GROCERY_DEMO_INSTANCE_URL'],
         environ['AITO_GROCERY_DEMO_API_KEY']
     )
 
-    suite = doctest.DocTestSuite(module=aito.client, extraglobs={'client': grocery_demo_client, 'pprint': pprint})
+    suite = doctest.DocTestSuite(module=aito.client.aito_client, extraglobs={'client': grocery_demo_client, 'pprint': pprint})
     tests.addTests(suite)
     return tests
